GE Debugger: Split playback into a separate file.
