+++ GPU/Debugger/Playback.h
@@ -17,23 +17,10 @@
 
 #pragma once
 
-#include <functional>
 #include <string>
-#include "Common/CommonTypes.h"
 
 namespace GPURecord {
 
-bool IsActive();
-bool IsActivePending();
-bool Activate();
-// Call only if Activate() returns true.
-void SetCallback(const std::function<void(const std::string &)> callback);
-
-void NotifyCommand(u32 pc);
-void NotifyMemcpy(u32 dest, u32 src, u32 sz);
-void NotifyMemset(u32 dest, int v, u32 sz);
-void NotifyUpload(u32 dest, u32 sz);
-void NotifyDisplay(u32 addr, int stride, int fmt);
-void NotifyFrame();
+bool RunMountedReplay(const std::string &filename);
 
 };
